+++ static/arguments.json
@@ -1742,6 +1742,26 @@
     ]
   },
   {
+    "component": "Markdown",
+    "status": "beta",
+    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/beta/markdown.rb",
+    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/beta/markdown/default/",
+    "parameters": [
+      {
+        "name": "tag",
+        "type": "Symbol",
+        "default": "`:div`",
+        "description": "One of `:article`, `:div`, or `:td`."
+      },
+      {
+        "name": "system_arguments",
+        "type": "Hash",
+        "default": "N/A",
+        "description": "[System arguments](/system-arguments)"
+      }
+    ]
+  },
+  {
     "component": "Popover",
     "status": "beta",
     "source": "https://github.com/primer/view_components/tree/main/app/components/primer/beta/popover.rb",
@@ -2206,26 +2226,6 @@
         "type": "Symbol",
         "default": "`:short`",
         "description": "One of `:long` or `:short`."
-      },
-      {
-        "name": "system_arguments",
-        "type": "Hash",
-        "default": "N/A",
-        "description": "[System arguments](/system-arguments)"
-      }
-    ]
-  },
-  {
-    "component": "Markdown",
-    "status": "beta",
-    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/markdown.rb",
-    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/markdown/default/",
-    "parameters": [
-      {
-        "name": "tag",
-        "type": "Symbol",
-        "default": "`:div`",
-        "description": "One of `:article`, `:div`, or `:td`."
       },
       {
         "name": "system_arguments",
