+++ static/audited_at.json
@@ -54,6 +54,7 @@
   "Primer::Beta::IconButton": "",
   "Primer::Beta::Label": "",
   "Primer::Beta::Link": "",
+  "Primer::Beta::Markdown": "",
   "Primer::Beta::Popover": "",
   "Primer::Beta::ProgressBar": "",
   "Primer::Beta::RelativeTime": "",
