+++ static/constants.json
@@ -709,6 +709,14 @@
       "span"
     ]
   },
+  "Primer::Beta::Markdown": {
+    "DEFAULT_TAG": "div",
+    "TAG_OPTIONS": [
+      "div",
+      "article",
+      "td"
+    ]
+  },
   "Primer::Beta::Popover": {
     "CARET_DEFAULT": "top",
     "CARET_MAPPINGS": {
@@ -987,12 +995,6 @@
     ]
   },
   "Primer::Markdown": {
-    "DEFAULT_TAG": "div",
-    "TAG_OPTIONS": [
-      "div",
-      "article",
-      "td"
-    ]
   },
   "Primer::MenuComponent": {
     "HEADING_TAG_FALLBACK": "h2",
