+++ static/statuses.json
@@ -54,6 +54,7 @@
   "Primer::Beta::IconButton": "beta",
   "Primer::Beta::Label": "beta",
   "Primer::Beta::Link": "beta",
+  "Primer::Beta::Markdown": "beta",
   "Primer::Beta::Popover": "beta",
   "Primer::Beta::ProgressBar": "beta",
   "Primer::Beta::RelativeTime": "beta",
@@ -77,7 +78,7 @@
   "Primer::LayoutComponent": "alpha",
   "Primer::LinkComponent": "deprecated",
   "Primer::LocalTime": "alpha",
-  "Primer::Markdown": "beta",
+  "Primer::Markdown": "deprecated",
   "Primer::MenuComponent": "alpha",
   "Primer::Navigation::TabComponent": "alpha",
   "Primer::OcticonComponent": "beta",
