+++ test/components/component_test.rb
@@ -84,7 +84,7 @@ class PrimerComponentTest < Minitest::Test
     [Primer::Beta::Label, {}],
     [Primer::LayoutComponent, {}],
     [Primer::Beta::Link, { href: "https://www.google.com" }],
-    [Primer::Markdown, {}],
+    [Primer::Beta::Markdown, {}],
     [Primer::MenuComponent, {}, proc { |c| c.item(href: "#url") { "Item" } }],
     [Primer::Navigation::TabComponent, {}],
     [Primer::OcticonComponent, { icon: :people }],
@@ -111,6 +111,7 @@ class PrimerComponentTest < Minitest::Test
 
   def test_registered_components
     ignored_components = [
+      "Primer::Markdown",
       "Primer::ClipboardCopy",
       "Primer::LabelComponent",
       "Primer::LinkComponent",
